+++ src/game/level.c
@@ -839,6 +839,7 @@ finish:
 
 bool __cdecl Level_Load(const char *const file_name, const int32_t level_num)
 {
+    LOG_DEBUG("%s (num=%d)", g_GF_LevelNames[level_num], level_num);
     init_game_malloc();
 
     BENCHMARK *const benchmark = Benchmark_Start();